requirements.txt
tensorflow
tensorflow-hub
numpy
paho-mqtt
requests
